Some refactoring and deletion of MainActivity since it's an unnecessary step
