+++ app/src/main/res/layout/activity_dashboard.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.example.badhri.huddle.activities.DashBoard">
+    tools:context="com.example.badhri.huddle.activities.DashboardActivity">
 
     <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
